feat(fleet)!: Duration at the boundaries, f64 seconds in the reports (#218)
**This is the chunk's one wire-visible change.** Six report fields that
serialized as integer seconds now serialize as `f64` seconds, so
`serde_json` renders `30` as `30.0`. The numbers are unchanged; the *type*
is now one type.

Seconds were spelled three ways at once across this surface — `u64` in
`CutoverReport.window_s`, `RateReport.window_s`, `BudgetWindow.window_s`,
`ScoutReport.timeout_s`, `CallReport.timeout_s`, zenctl's
`GetReport.timeout_s` and `RetiredReport.window_s: Asked<u64>`; `f64` in
`FieldReport.window_s`, `ExpectReport.window_s`, `WhyReport.listened_s`,
`ObservationSummary.window_s`, `BenchReport.elapsed_s`, every `Window` in
`condition.rs`. A consumer reading `window_s` off two documents of the same
tool got two JSON types for one concept, and the renderers papered over it
with `as f64` at four call sites (now gone). `f64` is the side that can
express a 2.5-second window, which `--listen-for` already accepts.

At the API boundary the unit stops being implied at all: `run_cutover` and
`cutover::scope_note`/`retired::scope_note` take a `Duration` instead of a
bare `u64`, and `run_retired` takes `Option<Duration>` instead of
`Option<u64>` — the frontends convert their whole-second flags at their own
edge, which is where a flag's unit belongs.

Six pinned assertions move, each with a comment citing the unification:
render.rs's budget `window_s` (10 → 10.0) and the call/probe/get `timeout_s`
(5 → 5.0); report_contract.rs's `CallReport` envelope (5 → 5.0) and
`RetiredReport.window_s` (30 → 30.0). Human output is unchanged — Rust's
`Display` for an integral `f64` prints no point, so every "within 5s" and
"30s window" sentence reads exactly as before.

+++ zenctl/src/render/impls/calls.rs
@@ -105,7 +105,7 @@ impl Render for CallReport {
     fn scope(&self) -> Option<ObservedScope> {
         Some(ObservedScope {
             asked: vec![self.key.clone()],
-            window_s: Some(self.timeout_s as f64),
+            window_s: Some(self.timeout_s),
         })
     }
 }
